backend/requirements.txt
# Django Core
Django>=5.0,<6.0
django-environ>=0.11
django-filter>=23.0

# Django REST Framework + Auth
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3.0
drf-spectacular>=0.27

# Database + PostGIS
psycopg[binary]>=3.1
django-cors-headers>=4.3
djangorestframework-gis>=1.2

# Media Storage
cloudinary>=1.36
pillow>=10.1

# Code Quality & Linting
flake8>=7.0

# Celery + Redis (Async tasks)
celery[redis]>=5.3
redis>=5.0